QCNode QNN: Fix QCMisraCPP2023 related issues
1. Mainly fix QCMisraCPP2023 related issues.
2. Remove the context binary generation from shared library as
not required feature.
3. Update Launch and Run user guide.

+++ source/Infras/Memory/Utils/BinaryAllocator.cpp
@@ -83,7 +83,7 @@ QCStatus_e BinaryAllocator::Free( const QCBufferDescriptorBase_t &buffer )
         QC_ERROR( "buffer not allocated" );
         status = QC_STATUS_INVALID_BUF;
     }
-    else if ( pid != pBufferDescriptor->pid )
+    else if ( static_cast<uint64_t>( pid ) != pBufferDescriptor->pid )
     {
         QC_ERROR( "buffer not allocated by self, can't do free" );
         status = QC_STATUS_OUT_OF_BOUND;
